pharma_backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
Pillow>=10.0.0
djangorestframework-simplejwt>=5.3.0
